+++ src/modes/direct_select.js
@@ -1,20 +1,26 @@
 import * as Constants from '../constants';
 import { isCircle, getCircleCenter } from '../utils/circle_geojson';
-import distance from '../utils/distance';
+import { distance, bearing } from '../utils/geodesy';
 import createGeodesicGeojson from '../utils/create_geodesic_geojson';
 
 function patchDirectSelect(DirectSelect) {
   const DirectSelectPatched = { ...DirectSelect };
 
   DirectSelectPatched.dragVertex = function(state, e, delta) {
-    const point = [e.lngLat.lng, e.lngLat.lat];
     const geojson = state.feature.toGeoJSON();
 
     if (isCircle(geojson)) {
-      const center = getCircleCenter(geojson);
-      const radius = distance(center, point);
-      state.feature.setProperty(Constants.properties.CIRCLE_RADIUS, radius);
-      state.feature.changed();
+      if (state.selectedCoordPaths[0] === '0.1') {
+        const center = getCircleCenter(geojson);
+        const handle = [e.lngLat.lng, e.lngLat.lat];
+        const radius = distance(center, handle);
+        const handleBearing = bearing(center, handle);
+        state.feature.properties[Constants.properties.CIRCLE_RADIUS] = radius;
+        state.feature[Constants.properties.CIRCLE_HANDLE_BEARING] = handleBearing;
+        state.feature.changed();
+      } else {
+        DirectSelect.dragFeature.call(this, state, e, delta);
+      }
     } else {
       DirectSelect.dragVertex.call(this, state, e, delta);
     }