+++ src/utils/create_geodesic_circle.js
@@ -1,10 +1,12 @@
-import circleToPolygon from 'circle-to-polygon';
+import { destinationPoint } from './geodesy';
 
-function createGeodesicCircle(center, radius, steps) {
-  // circleToPolygon expects center in -180..180
-  const shiftedCenter = [((center[0] + 180) % 360 + 360) % 360 - 180, center[1]];
-  const shiftedPolygon = circleToPolygon(shiftedCenter, radius * 1000, steps);
-  const coordinates = shiftedPolygon.coordinates[0].map(x => [x[0] + (center[0] - shiftedCenter[0]), x[1]]);
+function createGeodesicCircle(center, radius, bearing, steps) {
+  const coordinates = [];
+  for (let i = 0; i < steps; ++i) {
+    coordinates.push(destinationPoint(center, radius, bearing + (360 * -i) / steps));
+  }
+  coordinates.push(coordinates[0]);
+
   return coordinates;
 }