+++ src/utils/create_geodesic_geojson.js
@@ -2,20 +2,11 @@ import createVertex from '@mapbox/mapbox-gl-draw/src/lib/create_vertex';
 import * as Constants from '../constants';
 import { isCircle, getCircleCenter, getCircleRadius } from './circle_geojson';
 import createGeodesicLine from './create_geodesic_line';
-import createGeodesicMidpoint from './create_geodesic_midpoint';
 import createGeodesicCircle from './create_geodesic_circle';
+import { midpoint, destinationPoint } from './geodesy';
 
 const STEPS = 32;
-
-function minBy(array, func) {
-  const min = Math.min(...array.map(func));
-  return array.find(item => func(item) === min);
-}
-
-function maxBy(array, func) {
-  const max = Math.max(...array.map(func));
-  return array.find(item => func(item) === max);
-}
+const HANDLE_BEARING = 45;
 
 function getCoordinate(coordinates, path) {
   const ids = path.split('.').map(x => parseInt(x, 10));
@@ -31,10 +22,11 @@ function createGeodesicGeojson(geojson, options) {
   const coordinates = geojson.geometry.coordinates;
 
   const featureId = properties.parent || properties.id;
-  const featureGeojson = options.ctx.store.get(featureId).toGeoJSON();
+  const feature = options.ctx.store.get(featureId);
+  const featureGeojson = feature.toGeoJSON();
 
   if (type === Constants.geojsonTypes.POINT) {
-    if ((properties.meta === Constants.meta.VERTEX || properties.meta === Constants.meta.MIDPOINT) && featureGeojson && isCircle(featureGeojson)) {
+    if ((properties.meta === Constants.meta.VERTEX || properties.meta === Constants.meta.MIDPOINT) && isCircle(featureGeojson)) {
       return []; // hide circle points, they are displayed in processCircle instead
     } else if (properties.meta === Constants.meta.MIDPOINT) {
       return processMidpoint(); // calculate geodesic midpoint
@@ -44,7 +36,7 @@ function createGeodesicGeojson(geojson, options) {
   } else if (type === Constants.geojsonTypes.LINE_STRING) {
     return processLine(); // calculate geodesic line
   } else if (type === Constants.geojsonTypes.POLYGON) {
-    if (featureGeojson && isCircle(featureGeojson)) {
+    if (isCircle(featureGeojson)) {
       return processCircle(); // calculate geodesic circle
     } else {
       return processPolygon(); // calculate geodesic polygon
@@ -70,7 +62,7 @@ function createGeodesicGeojson(geojson, options) {
 
     const startCoord = getCoordinate(featureGeojson.geometry.coordinates, startCoordPath);
     const endCoord = getCoordinate(featureGeojson.geometry.coordinates, endCoordPath);
-    const midCoord = createGeodesicMidpoint(startCoord, endCoord);
+    const midCoord = midpoint(startCoord, endCoord);
 
     const geodesicGeojson = {
       ...geojson,
@@ -116,7 +108,8 @@ function createGeodesicGeojson(geojson, options) {
   function processCircle() {
     const center = getCircleCenter(featureGeojson);
     const radius = getCircleRadius(featureGeojson);
-    const geodesicCoordinates = createGeodesicCircle(center, radius, options.steps * 4);
+    const handleBearing = feature[Constants.properties.CIRCLE_HANDLE_BEARING] || HANDLE_BEARING;
+    const geodesicCoordinates = createGeodesicCircle(center, radius, handleBearing, options.steps * 4);
     const geodesicGeojson = {
       ...geojson,
       geometry: {
@@ -127,12 +120,8 @@ function createGeodesicGeojson(geojson, options) {
 
     // circle handles
     if (properties.active === Constants.activeStates.ACTIVE) {
-      const points = [
-        maxBy(geodesicCoordinates, x => x[0]), // north
-        minBy(geodesicCoordinates, x => x[0]), // south
-        maxBy(geodesicCoordinates, x => x[1]), // east
-        minBy(geodesicCoordinates, x => x[1])  // west
-      ];
+      const handle = destinationPoint(center, radius, handleBearing);
+      const points = [center, handle];
       const vertices = points.map((point, i) => {
         return createVertex(properties.id, point, `0.${i}`, isSelectedPath(`0.${i}`));
       })