+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- add center point, change radius handles from NSEW points to movable point
 ## 2.0.0
 
 - updated patching method of original modes, to be compatible with other patches